+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Workflow DataFlow type added
 ### Fixed
 - Workflow status fixed
+- Zones added in BaseModel
+- update input are Partial
+
 ## 2.5.2 - 2023-08-07
 ### Added
 - order model & interface