+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.6.1 - 2023-08-26
 ### Fixed
 - Addressing Common and Infrastructure folders
+
 ## 2.6.0 - 2023-08-15
 
 ## 2.5.4 - 2023-08-12